mcp_server/requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
requests==2.31.0
fabric_reports_client==1.0.0
fastmcp==0.3.0

# Configuration management
pydantic-settings==2.1.0
PyYAML==6.0.1
python-dotenv==1.0.1

# Structured logging
python-json-logger==2.0.7
